crypto: iaa - Add iaa_crypto_enable module param
Add a module param named iaa_crypto_enable, allowing userpace to
specify when IAA configuration is complete and thus the driver can be
enabled.

Enabling iaa_crypto_enable allows wq resources to be pinned down only
when being actively used.

With this patch, in order for iaa_crypto to do any compression or
decompression using the IAA hardware, the following must be done from
the command-line after all the desired iaa wqs and devices have been
configured and enabled:

    # echo 1 > /sys/module/iaa_crypto/parameters/iaa_crypto_enable

In order for the same iaa wqs and devices to be disabled, the
following must be done:

    # echo 0 > /sys/module/iaa_crypto/parameters/iaa_crypto_enable

Enabling or disabling these will display a corresponding message in
the kernel log:

    [  393.373502] iaa_crypto: iaa_crypto_enable: iaa_crypto now ENABLED

or

    [  562.694727] iaa_crypto: iaa_crypto_enable: iaa_crypto now DISABLED

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>