crypto: iaa - Add support for iaa_crypto async deflate compression algorithm
Add support for the iax_crypto async deflate algorithm, so that the
IAA crypto hardware can be used with facilities that require it, such
as zswap.

[NOTE: This currently only implements the async support needed for
zswap, which is the only facility that currently uses the async
interface but wraps it inside a sync wrapper anyway. This will support
async momentarily...]

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>