x86/virt/tdx: Invoke P-SEAMLDR APIs to load a TDX module
P-SEAMLDR provides a API (SEAMLDR_INSTALL) to allow VMM to install
a TDX module into MODULE_RANGE of SEAM range. The API should be called on
all logical processors in the platform, serially. When invoked on the first
logical processor, it starts the "installation session". When invoked on
the last logical processor, it loads or updates the MODULE_RANGE with the
given TDX module, and ends the "installation session".

If some CPUs don't call the API, the "installation session" will never
complete. So, kernel needs to ensure all CPUs are online and invoke the
API on all of them.

SEAMLDR_INSTALL may return the error, P_SEAMCALL_NO_ENTROPY, when P-SEAMLDR
cannot generate random number. This error is recoverable, kernel can simply
retry several times.

Also add a variation of __tdx_init() for callers that have already disabled
cpu hotplug and use it to initialize newly loaded TDX module.

Signed-off-by: Chao Gao <chao.gao@intel.com>