x86/topology: add support for CPUID.1F MODULE_TYPE
This is just a prototye patch, and it is not targeted for upstream.

How to fix the problem is still under discussion internally.

Teach the kernel about CPUID.1F MODULE_TYPE.

Before this update, the unrecognized MODULE_TYPE bits in the APIC-ID were
assumed to be package_id bits.  The result is that the kernel would
erroneously report a multi Ecore module system as a multi-package system.

With this update, the module_id bits are treated as an extension of the
core_id bits, so that we can get the correct package_id, as well as a
unique core_id within a package.

We don't currently see the need to complicate the topology namespace.
(eg. making core_id=7 into module_id=1 + core_id=3).

Note that with SCHED_CLUSTER, sysfs already identifies Ecore modules with
a unique cluster_id.

The patch has been tested on an 8 Atom core ADL-N platform.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Reported-and-tested-by: Wang Wendy <wendy.wang@intel.com>
Reviewed-by: Len Brown <len.brown@intel.com>