dmaengine: idxd: add per file user counters for completion record faults
Add counters per opened file for the char device in order to keep track how
many completion record faults occured and how many of those faults failed
the writeback by the driver after attempt to fault in the page. An xarray
is added to associate the pasid with the struct idxd_user_context so the
counters can be managed.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>