sched: Add interfaces for classes of tasks
On hybrid processors, the number of instructions retired per clock cycle
(IPC) is different among types of CPUs. Moreover, IPCs may even differ
among groups of instructions. The scheduler may improve the placement of
tasks if these differences are considered. It can, for instance, place
classes of tasks with specific IPC ratings on specific CPUs.

Add interfaces that architectures can implement to convey the needed
performance and classification to the scheduler.

Performance differences are indicated in the form of priorities, where
higher priority numbers indicate higher IPC for classification.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>