dmaengine: idxd: add new wq state for virtual device support
When a dedicated wq is enabled to support virtual device, the driver must
disable the wq on the device in order to program the PASID to the wq.
Introduce a wq state IDXD_WQ_LOCKED that is software state only in order to
prevent the user from modifying the configuration while the wq is in this
state. While in this state, the wq is not in DISABLED state and will
prevent any modifications to the configuration. It is also not in the
ENABLED state and therefore prevents any actions allowed in the NEABLED
state.

For the vfio_device, the dwq is disabled and set to LOCKED state on
vdev->open() call. The wq is changed to ENABLED state when the guest enables
the wq.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>