sched/core: Update the classification of task during scheduler tick
When supported, hardware continuously classify the current task. Thus, the
scheduler must periodically read the latest classification value. Reading
the latest classification value in the scheduler tick ensures that the
scheduler has an updated value when it may need it (when balancing load)
and that the classification is not read too frequently (a potentially
expensive operation).

Only update the classification when the tick is from user mode. This
ensures that the system has not spent too much time in kernel mode and
the classification ID that the hardware provides reflects better the
properties of the classified task.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>