thermal: intel: hfi: Report per-cpu class-specific performance scores
The interface arch_get_task_classid_prio() returns the priority of a
classification ID running on a CPU relative to the rest of the CPUs in the
system. Priority reflects differences on the number of instructions
retired per cycle on one type of CPU with respect to other types. Higher
priorities indicate a higher number of retired instructions retired.

The HFI table with support for Intel Thread director provides such data.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>