perf/x86: Support LBR event logging
The LBR event logging introduces a per-counter indication of precise
event occurrences in LBRs. It can provides a means to attribute exposed
retirement latency to combinations of events across a block of
instructions. It also provides a means of attributing Timed LBR
latencies to events.

Add a new event/group modifier, branch_events, to indicate the event for
LBR event logging.

The feature is only supported on the first 4 GP counters. Force the
event constraint to the first 4 GP counters.

Extend the struct perf_branch_entry to save the event logging.
For a group with branch_events, the four saturate values will be
returned in events order.

Only support one active group with branch events.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>