platform/x86: ISST: Use package numa infomation
When multiple numa nodes are in package, there can be only one match
in SST PCI device numa node as there is one device per package. In
this case match information from the other nodes in the same package.
This is fine as all the CPUs will share the same PCI device on that
package.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>