sched/fair: Collect load-balancing stats on the classes of tasks
When classes of tasks are supported, balancing load involves deciding
whether tasks need to be swapped between CPUs to increase throughput.

Determine what would be the score of each class of tasks in a scheduling
group if placedon the destination CPU. Only register those classes that
have higher scores on the destination CPU.

Register also the scores of the classes of tasks of the local group.
This data will be used when comparing the local group with a candidate
busiest group.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>