perf/x86: Add Intel Meteor Lake support
The Meteor Lake is similar to the Alder Lake. Both are hybrid
platforms, with e-core and p-core.

The key differences include:
- The e-core supports 2 PDIST counters (GP0 & GP1)
- New MSRs for OCR events are introduced for the e-core.
- New Data Source fields are introduced for the e-core.
- Retire Latency (Support in a separate patch) for both cores.
- LBR event logging (Support in a separate patch) for both cores.

Add new specific functions of extra_regs, data_source and
get_event_constraints to support new features on e-core.

The Retire Latency and LBR event logging are implemented with separate
patches.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>