x86/cpu: Disable kernel LASS when patching kernel          alternatives
Most of the kernel is mapped at virtual addresses
in the upper half of the address range. But kernel
deliberately initialized a temporary mm area
within the lower half of the address range
(a positive address range) for text poking, see
commit b0cdd87e0e93 ("x86/alternatives: Initialize
temporary mm for patching").

LASS stops access to a positive address (the lower
half address range) in kernel and this can be
deactivated if AC bit in EFLAGS register is set.
Hence using stac/clac instruction around access to
the address to avoid triggering a LASS #GP fault.

Signed-off-by: Yian Chen <yian.chen@intel.com>