lib/spdm: Do key cleanup
The keyring and temporary leaf keys were not disposed of in
get_certificates, leaving them registered, particularly after
error. Ensure leaf key is disposed if challenge fails. Note
that the key is still around if challenge succeeds. Clean up
in spdm_finish. This is so that it may be used by
get_measurements (TBA).

Signed-off-by: David E. Box <david.e.box@linux.intel.com>