lib/spdm: Add GET_MEASUREMENTS support
Add support for doing GET_MEASUREMENTS. Per the spec, measurement
requests can happen multiple times. If signing is supported a transcript
of the measurement requests will be taken until a final signing request
is made, at which point the transcript will be hashed according to the
algorithms selected after spdm_authorize has successfully be ran. The hash
will be used to verify the signature returned by the device (Responder).
The leaf key from spdm_authorize is used to verify the signature. Per the
spec this must be the from the first certificate chain. It may be better
to use a keychain for this purpose so that spdm_authorize doesn't need to
be reran on the first cert chain just to reacquire the leaf key. Uses a
callback to return the measurement data back to the caller.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>