platform/x86/intel/sdsi: Swap attestation mailbox command
SDSi mailbox messages must be 8-byte aligned. Because SPDM messages don't
have this requirement it is necessary to add padding. However, if padding
is added, ocode has to know what the unpadded size is. The original design
was to have the unpadded size appended add after the packet (after the
mailbox command). With this design the packet size didn't take into account
the unpadded SPDM field since this came after the mailbox command. The
design was changed to instead add the unpadded qword after the SPDM message
and have the mailbox command as the last qword after that. This allows the
packet size to now include the unpadded size field. It also keeps the
attestation mailbox design consistent with how the other mailbox commands
work.

Note, for testing in the QEMu model, this change did not require a QEMu
change since that code always looks for the mailbox command to be at the
end of the packet according to the packet_size.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>