KVM: TDX: retry seamcall on recoverable errors
Errors returned by TDX module with the non-recoverable bit (62)
cleared hints that it is likely to be a recoverable error. For
example, TDX_OPERAND_BUSY may be returned when the seamcall fails
to acquire a TDX internal lock. KVM needs to retry the seamcall in
those recoverable error cases.

Add a wrapper of the __seamcall to retry on the recoverable errors.
For the following special cases, retry isn't performed in the wrapper,
the error is instead handed over to the caller to take some action
before retry:
- TDX_VCPU_ASSOCIATED: indicates that the vcpu has been associated to
  another pcpu. Caller needs to disassociate the vcpu before retry.
- TDX_VCPU_NOT_ASSOCIATED: indicates that the vcpu hasn't been
  associated to any pcpu. For tdh_vp_flush, it doesn't need to retry.
- TDX_INTERRUPTED_RESUMABLE: caller usually needs to set a "resume"
  flag before retry the seamcall.

We also don't want to retry the seamcall for unlimited number of times
in case anything logically goes wrong, so set a relative large limit,
TDX_SEAMCALL_RETRY_MAX to capture any unexpected errors.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>