perf record: Fix "--per-thread" option for hybrid machines
Commit b91e5492f9d7 ("perf record: Add a dummy event on hybrid systems to
collect metadata records") adds a dummy event on hybrid systems to fix the
symbol "unknown" issue when the workload is created in a P-core but runs
on an E-core. When "--per-thread" is enabled, the nr_cpus is reduced to 1,
 adding a dummy event is useless for this issue, and it will also cause
"failed to mmap with 22 (Invalid argument)". This patch stops adding dummy
events when the option "--per-thread" is enabled, then the option can work
on the hybrid machines.

Before:

 # ./perf record -e cycles:u --per-thread  sleep 1
 failed to mmap with 22 (Invalid argument)

After:

 # ./perf record -e cycles:u --per-thread  sleep 1
[ perf record: Woken up 1 times to write data ]
[ perf record: Captured and wrote 0.002 MB perf.data (6 samples) ]

Fixes: b91e5492f9d7 ("perf record: Add a dummy event on hybrid systems to collect metadata records")
Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>