KVM: TDX: support userspace to set/get migration info from KVM
KVM_TDX_SET_MIGRATION_INFO is added for userspace needs to pass
some information (e.g. vsock port) that's required for pre-migration
setup. Once the info the set to the related binding slot, the slot
status is updated to TDX_BINDING_SLOT_STATUS_PREMIG_WAIT, meaning
that it is ready and waiting for the pre-migration process to read.

KVM_TDX_GET_MIGRATION_INFO is added for userspace to read the status
of the corresponding binding slot. Userspace should not start the
migration process when the status is still waiting for pre-migration
to be done.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>