perf/core: Fix the same task check in perf_event_set_output
With the --per-thread option, perf record errors out when sampling with
a hardware event and a software event as below.

 $ perf record -e cycles,dummy --per-thread ls
 failed to mmap with 22 (Invalid argument)

The same task is sampled by the two events. The IOC_OUTPUT is invoked
to share the mmap memory of the task. In the perf_event_set_output(),
the ctx is used to check whether the two events are for the same task.
However, a hardware event and a software event are from different task
context. The check fails.

The task struct is stored in the event->hw.target for each per-thread
event. It can be used to determine whether two events are for the same
task.

The patch can also fix another issue reported months ago.
https://lore.kernel.org/all/92645262-D319-4068-9C44-2409EF44888E@gmail.com/
The event->ctx is not ready when the perf_event_set_output() is invoked
in the perf_event_open(), while the event->hw.target has been assigned at
the moment.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>