platform/x86/intel-uncore-freq: Add support for multiple power domains
An SoC can contain multiple power domains with individual or collection
of mesh partitions. This partition is called fabric cluster.

Certain type of meshes will need to run at same frequency, they will be
placed in the same fabric cluster. Benefit of fabric cluster is that it
offers us a salable mechanism to deal with partitioned fabrics in a SoC.

The current sysfs interface support control at package and die level.
This interface is extended to support multiple power domain and fabric
clusters.

Only on SoCs with multiple power domains, the directory name will have a
suffix with an unique instance number. For example:
package_00_die_00.00, package_00_die_00.01 etc.
Attributes in the directory identify power domain and fabric cluster
id. This interface is expandable even if some new level of granularity
is introduced. A new sysfs attribute can identify new level.

For example on a single package Granite Rapids:
$tree /sys/devices/system/cpu/intel_uncore_frequency/
/sys/devices/system/cpu/intel_uncore_frequency/
├── package_00_die_00.00
│   ├── current_freq_khz
│   ├── domain_id
│   ├── fabric_cluster_id
│   ├── initial_max_freq_khz
│   ├── initial_min_freq_khz
│   ├── max_freq_khz
│   └── min_freq_khz
├── package_00_die_00.01
│   ├── current_freq_khz
│   ├── domain_id
│   ├── fabric_cluster_id
│   ├── initial_max_freq_khz
│   ├── initial_min_freq_khz
│   ├── max_freq_khz
│   └── min_freq_khz
└── package_00_die_00.02
    ├── current_freq_khz
    ├── domain_id
    ├── fabric_cluster_id
    ├── initial_max_freq_khz
    ├── initial_min_freq_khz
    ├── max_freq_khz
    └── min_freq_khz
    └── min_freq_khz

The attribute for cluster id is "fabric_cluster_id" instead of just
"cluster_id" is to avoid confusion with usage of term clusters in
other part of the Linux kernel.

Choice of adding "." to separate instance is already used in other
kernel sysfs interfaces.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>