sched/topology: Introduce counters for classes of tasks
Count the number of tasks of each class that hardware supports. Do
this for each runqueue.

When balancing load to maximize throughput based on classes of tasks, it is
necessary know which classes of tasks are present in a runqueue. Instead
of traversing the whole runqueue looking for suitable tasks at each balance
of load, count the number running tasks of each class.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>