perf/x86/intel: Optimize the fixed counter access
All the fixed counters share a fixed control register. Current perf
reads and rewrites the fixed control register for each fixed counter
disable/enable, which is unnecessary.

The fixed control register is only accessed when the entire PMU is
disabled. We just need to update the change right before the entire PMU
is reenabled.

The read of the fixed control is not unnecessary. We can cache the value
in per CPU cpu_hw_events.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>