x86/microcode/intel: Check against CPU signature before saving microcode
When looking to replace an entry in the microcode cache, only look for
current cpu signature against entries in the cache. Current code takes
the sig/pf from the saved_mc, and that's only looking for the top level
signature in the microcode file, and ignores the extended signature. The
vetting for microcode match is already done during scan_microcode(), and
save_microcode() only needs to save one copy of the microcode in
for the current CPU in the list.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Tested-by: William Xie <william.xie@intel.com>