x86/microcode/intel: Remove wbinvd() that's no longer required
Some older processors such as Broadwell with unbalanced core count to LLC
slices had a bad interaction when updating microcode and the caches were
dirty resulting in #MC. The wbinvd() was added to mitigate that before
doing microcode updates.

With the new minimim version required any older microcode will not be
loaded, and new microcode has no requirement for the wbinvd, Hence removing
that slow poke.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>