x86/microcode/intel: Enumerate capabilities to discover Uniform Microcode Loading.
Microcode update on Intel CPUs are gaining some new features. Uniform
Microcode loading is one of them. This mechanism simplifies runtime
microcode loading.

When CPU supports this feature, loading microcode on one core propagates to
the entire package or the entire platform.

Support enumeration of the feature that is used in subsequent patches to
implement uniform microcode loading.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>