x86/microcode: Handle uniform microcode update.
During late-load, perform apply microcode only to CPU in scope. When
uniform microcode is enabled, the CPU will propagate to all CPUs in scope.

When the package scope indicates PACKAGE, then all threads in the package
will receive the update automatically. When the scope indicates PLATFORM,
threads in all sockets receive the update.

Microcode updates for Intel CPUs contains not just the CPU microcode, but for
several other parts of the SOC. Each component authenticates its own
portion of the image and applies internally. Some cases not all

When the authentication fails, the update will result in tainting the
kernel.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>