perf test: Add ADL-N support for perf test 24
ADL-N only has cpu_atom, it is different from ADL-S/ADL-P which has both
cpu_core and cpu_atom, so the test case is different. Add ADL-N support for
perf test 24, the default event is "cycles", for the hybrid system, it
needs to change to "cpu_core/cycles/" when cpu_core is available and
"cpu_atom/cycles/" when only cpu_atom is available.

Before:
 # ./perf test 24
 24: Number of exit events of a simple workload               : FAILED!

After:
 # ./perf test 24
 24: Number of exit events of a simple workload               : Ok

Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>