sched/fair: Collect task class statistics for non-idle load balance
Collect non-idle load balance statistics for the classes of tasks in a
scheduling group. The purpose of these statistics is to find tasks that
would run with higher scores on the destination CPU.

For each of runqueue, collect the score of the current task of the
current queue if it was placed on the destination CPU. Register the
maximum score found (as this would be the task, and runqueue that would
benefit the most).

The purpose of these statistics is to find tasks that would be good
candidates to swap the task on the busiest runqueue with that of the
destination CPU. Therefore, we also need to collect the score of the
task currently running on the destination CPU to determine if swapping
tasks makes sense.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>