kvm: x86: enqcmd: update PASID translation table in VMX root mode only.
PASID Translation Table for ENQCMD/ENQCMDS is a per-VM table shared by
all vcpus/VMCS, it's not part of the VMCS, but linked by a pointer in
the VMCS. Per SDM 24.11.4, "Software should ensure that each such data
structure is modified only when no logical processor with a current VMCS
that references it is in VMX non-root mode.".

This patch uses kvm_make_block_vmentry_request to make sure all
vcpus are in VMX root mode when modify the PASID Translation Table.

Signed-off-by: Wu Hao <hao.wu@intel.com>
Signed-off-by: Yi Sun <yi.y.sun@linux.intel.com>