KVM: TDX: temporarily handle private mmio page fault
The private mmio is not set as the private memory slot now, it may have
no page struct either. So now the only way to get the pfn for it is to
fallthrough to the standard patch, same as shared mmio.

Add the private mmio to the management of user private memory (UPM)
could be the right way to go?

Signed-off-by: Xu Yilun <yilun.xu@intel.com>