x86/tdx: refactor mmio range parsing from device report
This makes the mmio range parsing easier to understand. Device report
gives enough information to identiy an mmio range as MSIX table, PBA, or
normal, so we don't have to calculate the range offsets according to
info in pcie configuration space.

But later on we still need to do more verification against these
info read from pcie configuration space, by comparing the device report
with configuration info, to ensure host pass through the device in a
right way. So the msix info prints are not deleted.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>