x86/virt/tdx: Check if the new module supports TD-preserving
A successful TD-preseving update requires support from both the old and
new TDX module.

There is no architectural way to check if a module supports a feature
without loading it. Heuristically use the num_handoff_pages to determine
if TD-preserving is supported. Note that the num_handoff_pages is The
number of 4KB pages (minus 1) required by this TDX module its handoff
data. So, in theory, a TDX module with TD-preserving support can have 0
in num_handoff_pages. But, num_handoff_pages was non-zero (i.e.,
requires at least 2 pages) from the beginning and won't decrease in
the future. It is safe to assume the theoretical case won't happen.

Signed-off-by: Chao Gao <chao.gao@intel.com>