x86/virt/tdx: Improve the compatibility check about handoff version
TD-preserving update involves passing metadata from the old module
to the new module. The metadata is stored in several pages at the
base address of SEAM_RANGE. To maintain compatibility, a version is
associated to the metadata (a.k.a, handoff data).

Each TDX module has two attributes, min_update_hv and module_hv.
They are the mininal version of handoff data can consume and the
native handoff version of the module respectively.

When OS requests TDX module to generate handoff data, the requested
version should be within the module's [min_update_hv, module_hv].
Otherwise, TDX module rejects the request. Note that a module with
no_downgrade flag is an exception.

And a module can consume handoff data iff the version is within its
[min_update_hv, module_hv].

Currently, we simply request current module to generate a handoff data
with version equal to min_update_hv of the module to load. If it fails,
then we refuse to load the new module.

But the check is inaccurate. E.g., current module has min_update_hv=1 and
module_hv=2 while the new module to load has min_update_hv=0, module_hv=1.
In this case, current check would fail. But in fact, the update should
succeed.

To fix this issue, retrieve supported version range from current module
and compare it with the supported version range of the new module to load.
If there is no overlap, refuse to load the new module. Otherwise, choose
the highest version supported by both.

Signed-off-by: Chao Gao <chao.gao@intel.com>