KVM: TDX: fix vfio group & device trust setting
The group attribute will not be reset by QEMU for legacy shared
passthrough, so vfio group should clear trusted attribute itself on group
close.

The trusted attribute of vfio device should also be updated according
to the group trusted attribute.

This fixes the issue that device could not be shared passed through if it
was previously private passed through.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>